_notes/requirements.txt
anyio @ file:///croot/anyio_1745334642479/work
argon2-cffi @ file:///opt/conda/conda-bld/argon2-cffi_1645000214183/work
argon2-cffi-bindings @ file:///croot/argon2-cffi-bindings_1736182440035/work
asttokens @ file:///croot/asttokens_1743630435401/work
async-lru @ file:///croot/async-lru_1699554519285/work
attrs @ file:///croot/attrs_1734533101012/work
babel @ file:///croot/babel_1737454360933/work
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
beautifulsoup4 @ file:///croot/beautifulsoup4-split_1718029820055/work
bleach @ file:///croot/bleach_1732290411627/work
Brotli @ file:///croot/brotli-split_1736182456865/work
certifi @ file:///croot/certifi_1745939216646/work/certifi
cffi @ file:///croot/cffi_1736182485317/work
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
click==8.1.8
comm @ file:///croot/comm_1709322850197/work
debugpy @ file:///croot/debugpy_1736267418885/work
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
defusedxml @ file:///tmp/build/80754af9/defusedxml_1615228127516/work
exceptiongroup @ file:///croot/exceptiongroup_1706031385326/work
executing @ file:///opt/conda/conda-bld/executing_1646925071911/work
fastjsonschema @ file:///croot/python-fastjsonschema_1731939362158/work
h11 @ file:///croot/h11_1748442006460/work
httpcore @ file:///croot/httpcore_1748526048470/work
httpx @ file:///croot/httpx_1746747840559/work
idna @ file:///croot/idna_1714398848350/work
importlib_metadata @ file:///croot/importlib_metadata-suite_1732633488278/work
ipykernel @ file:///croot/ipykernel_1737660677549/work
ipython @ file:///croot/ipython_1694181358621/work
jedi @ file:///croot/jedi_1733987392413/work
Jinja2 @ file:///croot/jinja2_1741710844255/work
json5 @ file:///croot/json5_1730786798687/work
jsonschema @ file:///croot/jsonschema_1728486696720/work
jsonschema-specifications @ file:///croot/jsonschema-specifications_1699032386549/work
jupyter-events @ file:///croot/jupyter_events_1741184577592/work
jupyter-lsp @ file:///croot/jupyter-lsp-meta_1745827015155/work
jupyter_client @ file:///croot/jupyter_client_1737570961872/work
jupyter_core @ file:///croot/jupyter_core_1718818295206/work
jupyter_server @ file:///croot/jupyter_server_1741200034823/work
jupyter_server_terminals @ file:///croot/jupyter_server_terminals_1744706698694/work
jupyterlab @ file:///croot/jupyterlab_1737555423487/work
jupyterlab_pygments @ file:///croot/jupyterlab_pygments_1741124142640/work
jupyterlab_server @ file:///croot/jupyterlab_server_1725865349919/work
MarkupSafe @ file:///croot/markupsafe_1738584038848/work
matplotlib-inline @ file:///opt/conda/conda-bld/matplotlib-inline_1662014470464/work
mistune @ file:///croot/mistune_1741124011532/work
nbclient @ file:///croot/nbclient_1741123995822/work
nbconvert @ file:///croot/nbconvert-meta_1741184653315/work
nbformat @ file:///croot/nbformat_1728049424075/work
nest-asyncio @ file:///croot/nest-asyncio_1708532673751/work
notebook @ file:///croot/notebook_1749732987674/work
notebook_shim @ file:///croot/notebook-shim_1741707758683/work
overrides @ file:///croot/overrides_1699371140756/work
packaging @ file:///croot/packaging_1734472117206/work
pandocfilters @ file:///opt/conda/conda-bld/pandocfilters_1643405455980/work
panflute==2.3.1
parso @ file:///croot/parso_1733963305961/work
pexpect @ file:///croot/pexpect_1750072593878/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
platformdirs @ file:///croot/platformdirs_1744273042065/work
prometheus_client @ file:///croot/prometheus_client_1744271615306/work
prompt-toolkit @ file:///croot/prompt-toolkit_1704404351921/work
psutil @ file:///croot/psutil_1736367091698/work
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
Pygments @ file:///croot/pygments_1744664109463/work
PySocks @ file:///tmp/build/80754af9/pysocks_1605305812635/work
python-dateutil @ file:///croot/python-dateutil_1716495738603/work
python-json-logger @ file:///croot/python-json-logger_1734370021104/work
PyYAML @ file:///croot/pyyaml_1728657952215/work
pyzmq @ file:///croot/pyzmq_1734687138743/work
referencing @ file:///croot/referencing_1699012038513/work
requests @ file:///croot/requests_1750426329888/work
rfc3339-validator @ file:///croot/rfc3339-validator_1683077044675/work
rfc3986-validator @ file:///croot/rfc3986-validator_1683058983515/work
rpds-py @ file:///croot/rpds-py_1736541261634/work
Send2Trash @ file:///croot/send2trash_1736540790175/work
six @ file:///croot/six_1744271502820/work
sniffio @ file:///croot/sniffio_1705431295498/work
soupsieve @ file:///croot/soupsieve_1696347547217/work
stack-data @ file:///opt/conda/conda-bld/stack_data_1646927590127/work
terminado @ file:///croot/terminado_1671751832461/work
tinycss2 @ file:///croot/tinycss2_1738337643607/work
tomli @ file:///opt/conda/conda-bld/tomli_1657175507142/work
tornado @ file:///croot/tornado_1748956929273/work
traitlets @ file:///croot/traitlets_1718227057033/work
typing_extensions @ file:///croot/typing_extensions_1734714854207/work
urllib3 @ file:///croot/urllib3_1737133630106/work
wcwidth @ file:///croot/wcwidth_1750352883074/work
webencodings==0.5.1
websocket-client @ file:///croot/websocket-client_1715878298792/work
zipp @ file:///croot/zipp_1732630741423/work
